Prevent Installing Dependencies in CI Scripts
Each CI script runs in its own isolated container, and therefore is not
bound to the dependency requirements of the codebase that it's running
for. The static analysis scripts run really slowly due to the fact that
they `bundle install` and `yarn install` before each task, and this is
not only unnecessary but actually breaks the build for some gems in
legacy stable versions. To address this, remove `bundle install` and
`yarn install` lines from each script, and install the gems/yarn
dependencies globally to the container. Although this removes any
plugin's ability to customize the RuboCop version, we can also be sure
that each plugin will be tested consistently and we won't be pinning
them to legacy older versions of our CI tools.